smartSplit/backend/requirements.txt
flask
python-dotenv
requests
langchain
langchain-google-genai